+++ chaosazure/postgresql/constants.py
@@ -1,0 +1,1 @@
+RES_TYPE_SRV_PG = "Microsoft.DBforPostgreSQL/servers"