+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 -  Introduce actions for Azure Netapp Files, like deleting netapp volumes
 -  Introduce actions and probes for Azure Storage, like deleting storage accounts and blob containers
+-  Introduce actions and probes for Azure Kubernetes Service, like deleting or stopping managed clusters
 
 ## [0.11.1][] - 2023-02-26