+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Added
 
 -  Introduce actions for Azure Netapp Files, like deleting netapp volumes
+-  Introduce actions and probes for Azure Storage, like deleting storage accounts and blob containers
 
 ## [0.11.1][] - 2023-02-26