+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 -  Introduce actions for Azure Netapp Files, like deleting netapp volumes
 -  Introduce actions and probes for Azure Storage, like deleting storage accounts and blob containers
 -  Introduce actions and probes for Azure Kubernetes Service, like deleting or stopping managed clusters
+-  Introduce action for Azure PostgreSQL Flexible to delete tables on one or many databases
 
 ## [0.11.1][] - 2023-02-26